requirements.txt
anthropic>=0.50.0
yt-dlp>=2024.12.0
python-dotenv>=1.0.0
rich>=13.0.0
aiofiles>=23.0.0
